Added simple terrain generation with water, grass, stone, dirt and craters

+++ MinecraftOpenGL/Source/Textures/TextureList.h
@@ -18,7 +18,7 @@ struct BlockTexture {
 
 	glm::vec2 textureAtlasOffset;
 
-	void SetTextureAtlasOffset(glm::ivec2 offset);
+	void SetTextureAtlasOffset(glm::vec2 offset);
 	void CalculateCoordinates();
 };
